refactor: Moved socket message handling to GrpcSocketProxy.
